refactor(impala): modernize the impala backend
Remove HDFS integration from the impala for dramatic simplification of the backend.

Also move away from an ancient version of impala that is no longer maintained.

The primary motivation here is to reduce the scope of the Impala backend, and to lower its maintenance by supporting fewer features.

BREAKING CHANGE: Direct HDFS integration is removed and support for ingesting pandas DataFrames directly is as well. The Impala backend still works with HDFS, but data in HDFS must be managed outside of ibis.

+++ ibis/backends/tests/test_dot_sql.py
@@ -61,8 +61,8 @@ def test_con_dot_sql(backend, con, schema):
         con.sql(
             f"""
             SELECT
-                string_col as s,
-                double_col + 1.0 AS new_col
+              string_col as s,
+              double_col + 1.0 AS new_col
             FROM {name}
             """,
             schema=schema,
@@ -84,7 +84,7 @@ def test_con_dot_sql(backend, con, schema):
         .sort_values()
         .reset_index(drop=True)
     )
-    backend.assert_series_equal(result, expected)
+    backend.assert_series_equal(result.astype(expected.dtype), expected)
 
 
 @table_dot_sql_notimpl
